assignment2/requirements.txt
haversine==2.8.1
mysql-connector-python==8.0.33
tabulate==0.9.0
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2